\docType{package}
\name{postMCMCglmm}
\alias{postMCMCglmm}
\alias{postMCMCglmm-package}
\title{Post Estimation for MCMCglmm}
\description{
  Post Estimation for MCMCglmm
}
\details{
  postMCMCglmm is a simple package to support a proposal at
  Modern Modeling Methods 2013 It has extraction functions
  and prediction functions for MCMCglmm models to help get
  average marginal predicted probabilities from a probit
  model.

  \tabular{ll}{ Package: \tab postMCMCglmm\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab 2013-1-30\cr
  License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }
}
\examples{
\dontrun{
  # fill me in!!
}
}
\author{
  Joshua Wiley \email{jwiley.psych@gmail.com},

  Maintainer: Joshua Wiley \email{jwiley.psych@gmail.com}
}
\seealso{
  See \code{\link{fixef.MCMCglmm}},
}
\keyword{package}

