\name{print.postMCMCglmmRE}
\alias{print.postMCMCglmmRE}
\title{Print method for extracted random effects (as standard deviations) from a MCMCglmm object}
\usage{
  \method{print}{postMCMCglmmRE} (x, ...)
}
\arguments{
  \item{x}{An \code{postMCMCglmmRE} object}

  \item{\dots}{Not currently used.}
}
\value{
  Nothing. Prints to console.
}
\description{
  Print method for extracted random effects (as standard
  deviations) from a MCMCglmm object
}
\examples{
\dontrun{
  # make me!!!
}
}
\seealso{
  \code{\link{stdranef}}
}

